pyproject.toml
[project]
name = "deployfixpy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.2",
    "z3-solver>=4.15.1.0",
]
